1-first-agent/requirements.txt
asana==5.0.0
openai==1.10.0
python-dotenv==0.13.0